requirements.txt
Flask~=3.1.0
Werkzeug~=3.1.3
python-dotenv~=1.1.0
openai~=1.77.0
sqlalchemy~=2.0.41
pydub~=0.25.1
pydantic~=2.11.4
protobuf~=6.31.0
tavily-python